requirements/local.txt
Django==1.11.4
Pillow==4.2.1
requests==2.18.3
